Moved database date handling to shared server utilities
ref https://linear.app/ghost/issue/BER-3851

Gift delivery needs the same UTC database timestamp formatting as automations. Keeping the implementation automation-owned would duplicate a database concern, so the existing formatter and parser now live beside the shared database date codec with their timezone behavior preserved.
