+++ ghost/core/test/unit/server/lib/db-date.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
 import {spawn} from 'node:child_process';
 import {once} from 'node:events';
-import {fromDatabaseDate, toDatabaseDate} from '../../../../../core/server/services/automations/database-date';
+import {fromDatabaseDate, toDatabaseDate} from '../../../../core/server/lib/db-date';
 
 describe('database date utilities', function () {
     const timezones = [
@@ -19,7 +19,7 @@ describe('database date utilities', function () {
     const runInOtherTimezones = async (toRun: string) => {
         // JSON.stringify does a good job wrapping strings in quotes and escaping.
         const s = JSON.stringify;
-        const modulePath = require.resolve('../../../../../core/server/services/automations/database-date');
+        const modulePath = require.resolve('../../../../core/server/lib/db-date');
 
         await Promise.all(timezones.map(async ({tz, expectedNaive}) => {
             const source = `
